Change event listener to suscriber to fix language issue
