+++ src/Binovo/ElkarBackupBundle/Listener/LocaleListener.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * @copyright 2015 - Xabi Ezpeleta (xezpeleta@gmail.com)
+ * @license http://www.opensource.org/licenses/bsd-license.php New-BSD
+ */
+
+namespace Binovo\ElkarBackupBundle\Listener;
+
+use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\HttpKernel\KernelEvents;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+
+
+class LocaleListener implements EventSubscriberInterface
+{
+    private $defaultLocale;
+    private $securityContext;
+    private $em;
+
+    public function __construct($defaultLocale = 'en')
+    {
+        $this->defaultLocale = $defaultLocale;
+    }
+
+    public function onKernelRequest(GetResponseEvent $event)
+    {
+        $request = $event->getRequest();
+        if (!$request->hasPreviousSession()) {
+            return;
+        }
+
+        // try to see if the locale has been set as a _locale routing parameter
+        if ($locale = $request->attributes->get('_locale')) {
+            $request->getSession()->set('_locale', $locale);
+        } else {
+            // if no explicit locale has been set on this request, use one from the session
+            $request->setLocale($request->getSession()->get('_locale', $this->defaultLocale));
+        }
+    }
+
+    public static function getSubscribedEvents()
+    {
+        return array(
+            // must be registered before the default Locale listener
+            KernelEvents::REQUEST => array(array('onKernelRequest', 17)),
+        );
+    }
+}
